fix: 修复 pool idx 溢出、CRC 错误码、transaction tick 死锁、frag_lens 截断
- route_pool_free: idx 从 uint8_t 改为 uint16_t 防止大 pool 溢出
- 新增 ROUTE_ERR_CRC 错误码，区分 CRC 校验失败与参数错误
- transaction_tick: 收集超时条目后锁外统一回调，避免递归死锁
- frag_lens 从 uint8_t* 改为 uint16_t*，支持单片 >255 字节

+++ sources/common/route_pool.c
@@ -38,7 +38,7 @@ void route_pool_free(route_pool_t *pool, uint8_t *block) {
     if (block == NULL) return;
     ptrdiff_t offset = block - pool->storage;
     if (offset < 0) return;
-    uint8_t idx = (uint8_t)(offset / pool->block_size);
+    uint16_t idx = (uint16_t)(offset / pool->block_size);
     if (idx >= pool->block_count) return;
     uint8_t w = idx / 32;
     uint8_t bit = idx % 32;
